requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
et_xmlfile==2.0.0
gtfs-realtime-bindings==1.0.0
idna==3.10
jedi==0.19.2
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
protobuf==5.29.3
pudb==2024.1.3
Pygments==2.19.1
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
urwid==2.6.16
urwid_readline==0.15.1
wcwidth==0.2.13
